# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todo API",
    "description": "A simple Todo application API built with FastAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/groups": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Get Groups",
        "description": "Get all groups for the current user.\n\nArgs:\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    List[Group]: List of groups owned by the user.",
        "operationId": "get_groups_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  },
                  "type": "array",
                  "title": "Response Get Groups Groups Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Create Group",
        "description": "Create a new group.\n\nArgs:\n    group: Group data to create.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Group: The created group.\n\nRaises:\n    HTTPException: If user has reached the maximum number of groups.",
        "operationId": "create_group_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/groups/{groupId}": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Read Group",
        "description": "Get a specific group by ID.\n\nArgs:\n    groupId: ID of the group to retrieve.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Group: The requested group.\n\nRaises:\n    HTTPException: If group not found or user doesn't have access.",
        "operationId": "read_group_groups__groupId__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Groupid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "groups"
        ],
        "summary": "Update Group",
        "description": "Update an existing group.\n\nArgs:\n    groupId: ID of the group to update.\n    group_update: Updated group data.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    Group: The updated group.\n\nRaises:\n    HTTPException: If group not found or user is not the owner.",
        "operationId": "update_group_groups__groupId__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Groupid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "groups"
        ],
        "summary": "Delete Group",
        "description": "Delete a group.\n\nNote: Todos associated with this group will have their group_id set to NULL.\n\nArgs:\n    groupId: ID of the group to delete.\n    db: Database session.\n    current_user: Current authenticated user.\n\nReturns:\n    dict: Success message.\n\nRaises:\n    HTTPException: If group not found or user is not the owner.",
        "operationId": "delete_group_groups__groupId__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Groupid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Group Groups  Groupid  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Group": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "Group name"
          },
          "color": {
            "$ref": "#/components/schemas/GroupColor",
            "default": "blue"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "userId": {
            "type": "integer",
            "title": "Userid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "title": "Group",
        "description": "Schema for group API responses."
      },
      "GroupColor": {
        "type": "string",
        "enum": [
          "blue",
          "green",
          "red",
          "yellow",
          "purple",
          "pink",
          "orange",
          "teal",
          "indigo",
          "gray"
        ],
        "title": "GroupColor",
        "description": "Predefiniowane kolory dla grup - sp\u00f3jne z palet\u0105 UI."
      },
      "GroupCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name",
            "description": "Group name"
          },
          "color": {
            "$ref": "#/components/schemas/GroupColor",
            "default": "blue"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "GroupCreate",
        "description": "Schema for creating a new group."
      },
      "GroupUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "color": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroupColor"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "GroupUpdate",
        "description": "Schema for updating a group."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}